requirements.txt
PyYAML~=6.0
openpyxl~=3.0.10
requests
instaloader~=4.9.5